requirements.txt
aiogram==3.*
aiohttp
python-dotenv
matplotlib==3.8.2
requests